refactor(tracing): restructure adapters to avoid circular imports
