+++ nemoguardrails/tracing/adapters/__init__.py
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .tracer import InteractionLog, Tracer
+from .adapter_factory import adapter_factory
